test: 회원 유닛 테스트 구
